Closes #12276
Issue in cancellation of OPD Individual Bills was corrected. Now after cancelling ones with mutiple payments using "Same as Billed", create multiple payments with acurate results in opd income report.
But have to recheck cancellation of batch bills. Its not handled. Have to start a new issue if it is there.